Added a few more effects, changed times to be floats internally, and added a better way to debug test certain ones.

+++ LuckyPills/Effects/HumeShield.cs
@@ -4,7 +4,7 @@ internal record HumeShield : PillEffect {
 	protected override bool IsEnabled { get; } = true;
 	protected override string DisplayText { get; } = "You've been given some shield";
 
-	protected override void OnEnabled(Player player, int duration) {
+	protected override void OnEnabled(Player player, float duration) {
 		Logger.Debug($"{this.GetType().Name} {System.Reflection.MethodBase.GetCurrentMethod().Name}");
 		player.HumeShield += 100f;
 	}
